% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbeh-package.R
\docType{package}
\name{msigdbeh-package}
\alias{msigdbeh-package}
\title{msigdbeh: MSigDB Data Frame via ExperimentHub}
\description{
Provides the Molecular Signatures Database (MSigDB) gene sets in an R data frame. Serves as the data source for 'msigdbr'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/igordot/msigdbeh}
  \item Report bugs at \url{https://github.com/igordot/msigdbeh/issues}
}

}
\author{
\strong{Maintainer}: Igor Dolgalev \email{igor.dolgalev@nyumc.org} (\href{https://orcid.org/0000-0003-4451-126X}{ORCID})

}
\keyword{internal}
